Move ViewTitle to ViewRouter child component
